Allow the user to logout with a malformed or expired token [#1257] (#1305)
* Allow the user to logout with a malformed or expired token.

* Fix formatting.

* Fix test comment.

* Update changelog.

* Bump fideslib version to raise a 403 if the supplied token is malformed instead of a 500.

* Allow the root user to logout.

+++ src/fidesops/ops/api/v1/endpoints/user_endpoints.py
@@ -1,9 +1,15 @@
+import json
 import logging
+from typing import Optional
 
+import jose.exceptions
 from fastapi import Depends, HTTPException, Security
 from fideslib.cryptography.cryptographic_util import b64_str_to_str
+from fideslib.cryptography.schemas.jwt import JWE_PAYLOAD_CLIENT_ID
+from fideslib.exceptions import AuthenticationError
 from fideslib.models.client import ClientDetail
 from fideslib.models.fides_user import FidesUser
+from fideslib.oauth.oauth_util import extract_payload
 from fideslib.oauth.schemas.user import UserPasswordReset, UserResponse, UserUpdate
 from sqlalchemy.orm import Session
 from starlette.status import (
@@ -14,12 +20,21 @@ from starlette.status import (
 )
 
 from fidesops.ops.api import deps
+from fidesops.ops.api.deps import get_db
 from fidesops.ops.api.v1 import urn_registry as urls
-from fidesops.ops.api.v1.scope_registry import USER_PASSWORD_RESET, USER_UPDATE
+from fidesops.ops.api.v1.scope_registry import (
+    SCOPE_REGISTRY,
+    USER_PASSWORD_RESET,
+    USER_UPDATE,
+)
 from fidesops.ops.api.v1.urn_registry import V1_URL_PREFIX
 from fidesops.ops.core.config import config
 from fidesops.ops.util.api_router import APIRouter
-from fidesops.ops.util.oauth_util import get_current_user, verify_oauth_client
+from fidesops.ops.util.oauth_util import (
+    get_current_user,
+    oauth2_scheme,
+    verify_oauth_client,
+)
 
 logger = logging.getLogger(__name__)
 router = APIRouter(tags=["Users"], prefix=V1_URL_PREFIX)
@@ -98,19 +113,50 @@ def update_user_password(
     return current_user
 
 
+def logout_oauth_client(
+    authorization: str = Security(oauth2_scheme), db: Session = Depends(get_db)
+) -> Optional[ClientDetail]:
+    """
+    Streamlined oauth checks for logout.  Only raises an error if no authorization is supplied.
+    Otherwise, regardless if the token is malformed or expired, still return a 204.
+    Returns a client if we can extract one from the token.
+    """
+    if authorization is None:
+        raise AuthenticationError(detail="Authentication Failure")
+
+    try:
+        token_data = json.loads(
+            extract_payload(authorization, config.security.app_encryption_key)
+        )
+    except jose.exceptions.JWEParseError:
+        return None
+
+    client_id = token_data.get(JWE_PAYLOAD_CLIENT_ID)
+    if (
+        not client_id or client_id == config.security.oauth_root_client_id
+    ):  # The root client is not a persisted object
+        return None
+
+    client = ClientDetail.get(
+        db, object_id=client_id, config=config, scopes=SCOPE_REGISTRY
+    )
+
+    return client
+
+
 @router.post(
     urls.LOGOUT,
     status_code=HTTP_204_NO_CONTENT,
 )
 def user_logout(
     *,
-    client: ClientDetail = Security(
-        verify_oauth_client,
-        scopes=[],
+    client: Optional[ClientDetail] = Security(
+        logout_oauth_client,
     ),
     db: Session = Depends(deps.get_db),
 ) -> None:
-    """Logout the user by deleting its client"""
+    """Logout the user by deleting its client where applicable"""
 
     logger.info("Logging out user.")
-    client.delete(db)
+    if client:
+        client.delete(db)
